Moved main logic to a separate class, added javadoc
