+++ encryption-decryption/src/com/nbugaenco/encryptdecrypt/Parameter.java
@@ -1,12 +1,39 @@
 package com.nbugaenco.encryptdecrypt;
 
+import java.util.Arrays;
+
+/**
+ * Parameters that used in CLI args for encoding-decoding process
+ */
 public enum Parameter {
+    /**
+     * Can be {@code enc} or {@code dec}
+     */
     MODE("-mode"),
+    /**
+     * Must be an integer number
+     */
     KEY("-key"),
+    /**
+     * Can be {@code shift} ({@link ShiftEncoder}) or {@code unicode} ({@link UnicodeEncoder})
+     */
     ALG("-alg"),
+    /**
+     * Any {@link String}
+     */
     DATA("-data"),
+    /**
+     * File path from which data will be read
+     */
     IN("-in"),
-    OUT("-out");
+    /**
+     * File path to which data will be written
+     */
+    OUT("-out"),
+    /**
+     * Parameter for printing help message
+     */
+    HELP("-help");
 
     final String label;
 
@@ -15,11 +42,8 @@ public enum Parameter {
     }
 
     public static Parameter valueOfLabel(String label) {
-        for (Parameter e : values()) {
-            if (e.label.equals(label)) {
-                return e;
-            }
-        }
-        return null;
+        return Arrays.stream(values())
+                .filter(it -> it.label.equals(label))
+                .findFirst().orElse(null);
     }
 }
