+++ encryption-decryption/src/com/nbugaenco/encryptdecrypt/ShiftEncoder.java
@@ -1,22 +1,39 @@
 package com.nbugaenco.encryptdecrypt;
 
+/**
+ * This {@link Encoder} uses shift method.
+ * It's only changing english letters inside the alphabet.
+ * For example: {@code -key 5 -data "Hello World!"} will give {@code "Mjqqt Btwqi!"}
+ */
 public class ShiftEncoder extends Encoder {
 
     public ShiftEncoder() {
     }
 
+    private Integer encodeChar(Integer c) {
+        if (c >= 65 && c <= 90) {
+            c = (c + key > 90) ? (c + key) - 26 : c + key;
+        } else if (c >= 97 && c <= 122) {
+            c = (c + key > 122) ? (c + key) - 26 : c + key;
+        }
+
+        return c;
+    }
+
+    private Integer decodeChar(Integer c) {
+        if (c >= 65 && c <= 90) {
+            c = (c - key < 65) ? (c - key) + 26 : c - key;
+        } else if (c >= 97 && c <= 122) {
+            c = (c - key < 97) ? (c - key) + 26 : c - key;
+        }
+
+        return c;
+    }
+
     @Override
     public String encode() {
         return prompt.codePoints()
-                .map(c -> {
-                    if (c >= 65 && c <= 90) {
-                        c = (c + key > 90) ? (c + key) - 26 : c + key;
-                    } else if (c >= 97 && c <= 122) {
-                        c = (c + key > 122) ? (c + key) - 26 : c + key;
-                    }
-
-                    return c;
-                })
+                .map(this::encodeChar)
                 .collect(StringBuilder::new,
                         StringBuilder::appendCodePoint,
                         StringBuilder::append)
@@ -26,15 +43,7 @@ public class ShiftEncoder extends Encoder {
     @Override
     public String decode() {
         return prompt.codePoints()
-                .map(c -> {
-                    if (c >= 65 && c <= 90) {
-                        c = (c - key < 65) ? (c - key) + 26 : c - key;
-                    } else if (c >= 97 && c <= 122) {
-                        c = (c - key < 97) ? (c - key) + 26 : c - key;
-                    }
-
-                    return c;
-                })
+                .map(this::decodeChar)
                 .collect(StringBuilder::new,
                         StringBuilder::appendCodePoint,
                         StringBuilder::append)
