+++ encryption-decryption/src/com/nbugaenco/encryptdecrypt/UnicodeEncoder.java
@@ -1,5 +1,10 @@
 package com.nbugaenco.encryptdecrypt;
 
+/**
+ * This {@link Encoder} uses unicode method.
+ * It's changing all characters based on {@code -key} value
+ * For example: {@code -key 10 -alg unicode -data "Hello World!"} will give {@code "Rovvy*ay|vn+"}
+ */
 public class UnicodeEncoder extends Encoder {
 
     public UnicodeEncoder() {
